website/requirements/req_base.txt
bpython
Django==1.8.*
django-admin-honeypot==0.4.*
django-adminactions==0.8.*
django-extensions==1.5.*
django-suit==0.2.*
Pillow==2.9.*
psycopg2
python-postmark==0.4.*
raven==5.*
wagtail
